feat: composer configuration & remove /vendor
add workflow for composer validation
fix: composer install for php7 but update for php8

+++ install/restore.php
@@ -122,24 +122,25 @@ try {
 	system('cd ' . $jeedom_dir . '; tar xfz "' . $backup . '" ' . $exclude);
 	echo "OK\n";
 
-	
-	if (exec('which composer | wc -l') == 0) {
-		echo "\nNeed to install composer...";
-		echo shell_exec(system::getCmdSudo().' ' . __DIR__ . '/../resources/install_composer.sh');
-		echo "OK\n";
-	}
-	echo "Update composer file...\n";
-	if (exec('which composer | wc -l') > 0) {
-		shell_exec(system::getCmdSudo(). ' rm '. __DIR__ . '/../composer.lock');
-		shell_exec('export COMPOSER_HOME="/tmp/composer";export COMPOSER_ALLOW_SUPERUSER=1;'.system::getCmdSudo().' composer self-update > /dev/null 2>&1');
-		shell_exec('cd ' . __DIR__ . '/../;export COMPOSER_ALLOW_SUPERUSER=1;export COMPOSER_HOME="/tmp/composer";'.system::getCmdSudo().' composer update --no-interaction --no-plugins --no-scripts --no-ansi --no-dev --no-progress --optimize-autoloader --with-all-dependencies --no-cache > /dev/null 2>&1');
-		shell_exec(system::getCmdSudo().' rm /tmp/composer 2>/dev/null');
-		if(method_exists('jeedom','cleanFileSystemRight')){
-			jeedom::cleanFileSystemRight();
+	if(version_compare(PHP_VERSION, '8.0.0') >= 0){
+		if (exec('which composer | wc -l') == 0) {
+			echo "\nNeed to install composer...";
+			echo shell_exec(system::getCmdSudo().' ' . __DIR__ . '/../resources/install_composer.sh');
+			echo "OK\n";
 		}
+		
+		if (exec('which composer | wc -l') > 0) {
+			shell_exec('export COMPOSER_HOME="/tmp/composer";export COMPOSER_ALLOW_SUPERUSER=1;'.system::getCmdSudo().' composer self-update > /dev/null 2>&1');
+			shell_exec('cd ' . __DIR__ . '/../;export COMPOSER_ALLOW_SUPERUSER=1;export COMPOSER_HOME="/tmp/composer";'
+			  . system::getCmdSudo().' composer install --no-progress --no-interaction --no-dev > /dev/null 2>&1');
+			shell_exec(system::getCmdSudo().' rm /tmp/composer 2>/dev/null');
+			if(method_exists('jeedom','cleanFileSystemRight')){
+				jeedom::cleanFileSystemRight();
+			}
+		}
+		echo "OK\n";
+		echo "[PROGRESS][58]\n";
 	}
-	echo "OK\n";
-	echo "[PROGRESS][58]\n";
 
 	if (!file_exists($jeedom_dir . "/DB_backup.sql")) {
 		throw new Exception('Cannot find database backup file : DB_backup.sql');
@@ -167,11 +168,7 @@ try {
 			$str_db_connexion = "--host=" . $CONFIG['db']['host'] . " --port=" . $CONFIG['db']['port'] . " --user=" . $CONFIG['db']['username'] . " --password='" . $CONFIG['db']['password'] . "' " . $CONFIG['db']['dbname'];
 		}
 	}
-	if(isset($CONFIG['db']['unix_socket'])) {
-		shell_exec("mysql ". $str_db_connexion . "  < " . $jeedom_dir . "/DB_backup.sql");
-	} else {
-		shell_exec("mysql ". $str_db_connexion . "  < " . $jeedom_dir . "/DB_backup.sql");
-	}
+	shell_exec("mysql ". $str_db_connexion . "  < " . $jeedom_dir . "/DB_backup.sql");
 	echo "OK\n";
 	
 	echo "Enable back constraints...";
